whatsapp/requirements.txt
twilio
python-dotenv
fastapi[standard]
uvicorn
httpx